Refactor lib name: add namespace
