+++ device/vulkan/include/VulkanErrors.h
@@ -8,31 +8,36 @@
 #include <exception>
 #include <string>
 
-class InvalidEnum : public std::exception {
-public:
-    ~InvalidEnum() noexcept override = default;
-    const char *what() const noexcept override {
-        return "VulkanDefinitions: invalid input enum";
-    }
-};
+namespace ignimbrite {
 
-class VulkanException : public std::exception {
-public:
-    explicit VulkanException(const char* message) noexcept
-        : message(message) {
-    }
-    ~VulkanException() noexcept override = default;
-    const char *what() const noexcept override {
-        return message.c_str();
-    }
-private:
-    std::string message;
-};
+    class InvalidEnum : public std::exception {
+    public:
+        ~InvalidEnum() noexcept override = default;
+        const char *what() const noexcept override {
+            return "VulkanDefinitions: invalid input enum";
+        }
+    };
 
-inline void VulkanCheck(VkResult result) {
-    if (result != VK_SUCCESS) {
-        throw VulkanException("Operation failed");
+    class VulkanException : public std::exception {
+    public:
+        explicit VulkanException(const char *message) noexcept
+                : message(message) {
+        }
+        ~VulkanException() noexcept override = default;
+        const char *what() const noexcept override {
+            return message.c_str();
+        }
+
+    private:
+        std::string message;
+    };
+
+    inline void VulkanCheck(VkResult result) {
+        if (result != VK_SUCCESS) {
+            throw VulkanException("Operation failed");
+        }
     }
-}
+
+} // namespace ignimbrite
 
 #endif //RENDERINGLIBRARY_VULKANERRORS_H
