+++ ignimbrite/include/ignimbrite/FileUtils.h
@@ -8,9 +8,13 @@
 #include <vector>
 #include <string>
 
-class FileUtils {
-public:
-    static void loadData(const std::string &filename, std::vector<char> &data);
-};
+namespace ignimbrite {
+
+    class FileUtils {
+    public:
+        static void loadData(const std::string &filename, std::vector<char> &data);
+    };
+
+} // namespace ignimbrite
 
 #endif //VULKANRENDERER_FILEUTILS_H
