+++ ignimbrite/include/ignimbrite/Options.h.in
