+++ ignimbrite/include/ignimbrite/Platform.h
